free/requirements.txt
flask==2.3.3
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
schedule==1.2.0
telegram-bot==0.1.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
